Fix structure of autocompletion service
